feat: refactor src libraries into auth and storage
