+++ src/auth.js
@@ -1,11 +1,9 @@
 import { AppConfig, UserSession, showConnect } from '@stacks/auth';
 import { Person } from '@stacks/profile';
-import { Storage } from '@stacks/storage';
 
 const appConfig = new AppConfig(['store_write', 'publish_data']);
 
 export const userSession = new UserSession({ appConfig });
-export const storage = new Storage({ userSession });
 
 export function authenticate() {
   showConnect({
